added the need to be super admin to delete handlers

+++ src/api/teams/handlers/delete.ts
@@ -4,6 +4,11 @@ import { ITeamHandlers } from "./../interface";
 const deleteTeam: ITeamHandlers["delete"] = async (req, res) => {
   const { id } = req.params;
   try {
+    if (!req.user || req.user.role !== "SUPER_ADMIN") {
+      return res
+        .status(403)
+        .json({ message: "Forbidden, you don't have the right access" });
+    }
     const deleteATeam = await prisma.team.delete({
       where: { id },
     });
